bump token-gradle version; be more rigorous about `apply plugin`
Bump token-gradle version.

This hit a snag: We were trying to add an extension twice, an extension
added by io.token.gradle.lib , which we were applying in the root.
Changed to apply it at the leaves.

And that made me notice it was strange we were applying the version plugin
to everything in the sdk; just one project uses it.